+++ extras/TestCode.R
@@ -5,20 +5,6 @@ insertCohortDefinitionInPackage(definitionId = 5021,
 
 
 
-# Test mail notifications -------------------------------------------------
-mailSettings <- list(from = Sys.getenv("mailAddress"),
-                     to = c(Sys.getenv("mailAddress")),
-                     smtp = list(host.name = "smtp.gmail.com",
-                                 port = 465,
-                                 user.name = Sys.getenv("mailAddress"),
-                                 passwd = Sys.getenv("mailPassword"),
-                                 ssl = TRUE),
-                     authenticate = TRUE,
-                     send = TRUE)
-
-runAndNotify({
-  a <- b
-}, mailSettings = mailSettings, label = "Fancy code")
 
 
 
@@ -33,7 +19,10 @@ getPriorityVocabKey(baseUrl = Sys.getenv("baseUrl"))
 getConceptSetConceptIds(baseUrl = Sys.getenv("baseUrl"), setId = 12)
 
 
-# Args functions --------------------------------------------------------------
+# R environment snapshot ------------------------------------------------------
+# snapshot <- read.csv("c:/temp/rEnvironmentSnapshot.csv")
+# restoreEnvironment(snapshot)
+
+restoreEnvironmentFromPackageOnGithub("OHDSI/Legend")
 
 
-createArgFunction(functionName = "getCohortDefinitionName", rCode = "")