+++ NEWS.md
@@ -1,7 +1,19 @@
+OhdsiRTools v1.7.0
+==================
+
+Changes: 
+
+1. Restoring R environment now fetches OHDSI packages from drat repository. 
+
+2. Restoring R environment has option (default) to be less strict, only installing packages if installed version is older than required, or has different major version number.
+
+2. Added function for storing R environments from packages (including on GitHub)
+
+
 OhdsiRTools v1.6.0
 ==================
 
-Deprecating logging and parallel computation functions, which have been moved to ParallelLogger.
+Changes: Deprecating logging and parallel computation functions, which have been moved to ParallelLogger.
 
 
 OhdsiRTools v1.5.5